work tracer

+++ core/vm/noop.go
@@ -54,7 +54,7 @@ func (NoopStateDB) SetNonce(common.Address, uint64)                             
 func (NoopStateDB) GetCodeHash(common.Address) common.Hash                             { return common.Hash{} }
 func (NoopStateDB) GetCode(common.Address) []byte                                      { return nil }
 func (NoopStateDB) SetCode(common.Address, []byte)                                     {}
-func (NoopStateDB) GetCodeSize(common.Address) int                                     { return 0 }
+func (NoopStateDB) GetCodeSize(common.Address) uint64                                  { return 0 }
 func (NoopStateDB) AddRefund(uint64)                                                   {}
 func (NoopStateDB) GetRefund() uint64                                                  { return 0 }
 func (NoopStateDB) GetState(common.Address, common.Hash) common.Hash                   { return common.Hash{} }
